+++ firmware/hdl/packages/lhc_data_pkg.vhd
@@ -2,7 +2,8 @@
 -- Package for definitions of lhc data.
 
 -- Version history:
--- HB 2023-06-23: inserted calo anomaly algorithm (CICADA) data on link 11 (removed zdc10g on link 11).
+-- HB 2023-09-06: inserted calo anomaly algorithm (CICADA) data on link 11 (removed zdc10g on link 11).
+-- HB 2023-08-25: ZDC 5G on link 71.
 -- HB 2023-02-09: updated comments.
 -- HB 2022-10-08: zdc10g on link 11.
 -- HB 2022-09-08: cleaned up.
@@ -36,6 +37,7 @@
 -- muon                2 (3..2)        1        1
 -- muon                2 (5..4)        1        2
 -- muon                2 (7..6)        1        3
+-- zdc                 6               1       71
 -- __________________________________________________________
 -- Summary             72 (32 bits)    16
 --
@@ -57,8 +59,9 @@
 -- muon        3h     3l    2h   2l   x      x
 -- muon        5h     5l    4h   4l   x      x
 -- muon        7h     7l    6h   6l   x      x
+-- zdc (5G)    5      4     3    2    1      0     
 --
--- *) in addition 4 MSBs for Minimum Bias bits. TC means "TOWERCOUNT" (ECAL sum)
+-- *) in addition 4 MSBs for Minimum Bias bits. TC means "TOWERCNT" (ECAL sum)
 
 --===============================================================================================--
 
@@ -76,12 +79,10 @@ package lhc_data_pkg is
     constant OFFSET_TAU_LANES : natural := 8;
     constant OFFSET_ESUMS_LANES : natural := 10;
 -- HB 2106-05-31: proposal for memory structure with all frames of calo links for extended test-vector-file structure (see lhc_data_pkg_all_frames.vhd)
-    constant OFFSET_ZDC10G_LANES : natural := 11;
+    constant OFFSET_LINK_11_LANES : natural := 11;
     constant OFFSET_EXT_COND_LANES : natural := 12;
     constant OFFSET_CICADA_LANES : natural := 13;
-
--- HB 2022-10-10: lane number of ZDC 5G optical input
-    constant ZDC5G_LANE_NR : natural := 71;
+    constant OFFSET_ZDC5G_LANES : natural := 71;
 
 --  for simspy memory (test with ipb_dpmem_4096_32)
     constant SW_DATA_WIDTH : integer := 32;
@@ -103,6 +104,12 @@ package lhc_data_pkg is
     constant CICADA_ARRAY_LENGTH : integer := 6;
     constant CICADA_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant EXTERNAL_CONDITIONS_DATA_WIDTH : integer := SW_DATA_WIDTH*8;
+    constant ZDC5G_0_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_1_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_2_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_3_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_4_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_5_WIDTH : integer := SW_DATA_WIDTH;
 
     constant LHC_DATA_WIDTH : integer :=
         (
@@ -118,6 +125,10 @@ package lhc_data_pkg is
             HTMHF_DATA_WIDTH +
             (CICADA_ARRAY_LENGTH*CICADA_DATA_WIDTH) +
             EXTERNAL_CONDITIONS_DATA_WIDTH
+            EXTERNAL_CONDITIONS_DATA_WIDTH +
+            ZDC5G_0_WIDTH + ZDC5G_1_WIDTH +
+            ZDC5G_2_WIDTH + ZDC5G_3_WIDTH +
+            ZDC5G_4_WIDTH + ZDC5G_5_WIDTH
         );
 
     type muon_array_t is array(0 to MUON_ARRAY_LENGTH-1) of std_logic_vector(MUON_DATA_WIDTH-1 downto 0);
@@ -139,6 +150,12 @@ package lhc_data_pkg is
         htmhf : std_logic_vector(HTMHF_DATA_WIDTH-1 downto 0);
         cicada : cicada_array_t;
         external_conditions : std_logic_vector(EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto 0);
+        ZDC5G_0 : std_logic_vector(ZDC5G_0_WIDTH-1 downto 0);
+        ZDC5G_1 : std_logic_vector(ZDC5G_1_WIDTH-1 downto 0);
+        ZDC5G_2 : std_logic_vector(ZDC5G_2_WIDTH-1 downto 0);
+        ZDC5G_3 : std_logic_vector(ZDC5G_3_WIDTH-1 downto 0);
+        ZDC5G_4 : std_logic_vector(ZDC5G_4_WIDTH-1 downto 0);
+        ZDC5G_5 : std_logic_vector(ZDC5G_5_WIDTH-1 downto 0);
     end record;
 
     constant LHC_DATA_NULL : lhc_data_t :=
@@ -154,10 +171,16 @@ package lhc_data_pkg is
             etmhf => (others=>'0'),
             htmhf => (others=>'0'),
             cicada => (others=>(others=>'0')),
-            external_conditions => (others=>'0')
+            external_conditions => (others=>'0'),
+            ZDC5G_0 => (others=>'0'),
+            ZDC5G_1 => (others=>'0'),
+            ZDC5G_2 => (others=>'0'),
+            ZDC5G_3 => (others=>'0'),
+            ZDC5G_4 => (others=>'0'),
+            ZDC5G_5 => (others=>'0')
         );
 
-    constant LHC_DATA_OBJECT_COUNT : integer :=12;
+    constant LHC_DATA_OBJECT_COUNT : integer :=18;
     constant INDEX_MUON : integer := 0;
     constant INDEX_EG : integer := 1;
     constant INDEX_TAU : integer := 2;
@@ -170,6 +193,12 @@ package lhc_data_pkg is
     constant INDEX_HTMHF : integer := 9;
     constant INDEX_CICADA : integer := 10;
     constant INDEX_EXTERNAL_CONDITIONS : integer := 11;
+    constant INDEX_ZDC5G_0 : integer := 12;
+    constant INDEX_ZDC5G_1 : integer := 13;
+    constant INDEX_ZDC5G_2 : integer := 14;
+    constant INDEX_ZDC5G_3 : integer := 15;
+    constant INDEX_ZDC5G_4 : integer := 16;
+    constant INDEX_ZDC5G_5 : integer := 17;
     
     type lhc_data_slv_property_t is array (0 to LHC_DATA_OBJECT_COUNT-1) of natural;
 
@@ -187,6 +216,9 @@ package lhc_data_pkg is
             HTMHF_DATA_WIDTH,
             CICADA_ARRAY_LENGTH * CICADA_DATA_WIDTH,         
             EXTERNAL_CONDITIONS_DATA_WIDTH
+            ZDC5G_0_WIDTH, ZDC5G_1_WIDTH,
+            ZDC5G_2_WIDTH, ZDC5G_3_WIDTH,
+            ZDC5G_4_WIDTH, ZDC5G_5_WIDTH
         );
 
     constant LHC_DATA_SLV_START_INDICES : lhc_data_slv_property_t; -- see body
@@ -249,6 +281,19 @@ package body lhc_data_pkg is
         ret_value(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index) := data_in.external_conditions;
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
 
+        ret_value(index + ZDC5G_0_WIDTH-1 downto index) := data_in.ZDC5G_0;
+        index := index + ZDC5G_0_WIDTH;
+        ret_value(index + ZDC5G_1_WIDTH-1 downto index) := data_in.ZDC5G_1;
+        index := index + ZDC5G_1_WIDTH;
+        ret_value(index + ZDC5G_2_WIDTH-1 downto index) := data_in.ZDC5G_2;
+        index := index + ZDC5G_2_WIDTH;
+        ret_value(index + ZDC5G_3_WIDTH-1 downto index) := data_in.ZDC5G_3;
+        index := index + ZDC5G_3_WIDTH;
+        ret_value(index + ZDC5G_4_WIDTH-1 downto index) := data_in.ZDC5G_4;
+        index := index + ZDC5G_4_WIDTH;
+        ret_value(index + ZDC5G_5_WIDTH-1 downto index) := data_in.ZDC5G_5;
+        index := index + ZDC5G_5_WIDTH;
+
         return ret_value;
     end function;
 
@@ -298,6 +343,19 @@ package body lhc_data_pkg is
 
         ret_value.external_conditions := data_in(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index);
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
+
+        ret_value.ZDC5G_0 := data_in(index + ZDC5G_0_WIDTH-1 downto index);
+        index := index + ZDC5G_0_WIDTH;
+        ret_value.ZDC5G_1 := data_in(index + ZDC5G_1_WIDTH-1 downto index);
+        index := index + ZDC5G_1_WIDTH;
+        ret_value.ZDC5G_2 := data_in(index + ZDC5G_2_WIDTH-1 downto index);
+        index := index + ZDC5G_2_WIDTH;
+        ret_value.ZDC5G_3 := data_in(index + ZDC5G_3_WIDTH-1 downto index);
+        index := index + ZDC5G_3_WIDTH;
+        ret_value.ZDC5G_4 := data_in(index + ZDC5G_4_WIDTH-1 downto index);
+        index := index + ZDC5G_4_WIDTH;
+        ret_value.ZDC5G_5 := data_in(index + ZDC5G_5_WIDTH-1 downto index);
+        index := index + ZDC5G_5_WIDTH;
 
         return ret_value;
     end function;