+++ firmware/hdl/payload/gtl_data_mapping.vhd
@@ -2,7 +2,8 @@
 -- Data mapping for GTL
 
 -- Version-history:
--- HB 2023-06-23: cicada on link 11 (removed zdc10g on link 11).
+-- HB 2023-09-06: cicada on link 11 (removed zdc10g on link 11).
+-- HB 2023-08-25: updated for new ZDC data structure.
 -- HB 2023-03-01: updated for ZDC.
 -- HB 2023-02-09: updated comments.
 -- HB 2022-10-10: inserted ZDC.
@@ -95,6 +96,13 @@ begin
 
 -- ****************************************************************************************
 
+    gtl_data.zdc(0) <= lhc_data.zdc5g_0;
+    gtl_data.zdc(1) <= lhc_data.zdc5g_1;
+    gtl_data.zdc(2) <= lhc_data.zdc5g_2;
+    gtl_data.zdc(3) <= lhc_data.zdc5g_3;
+    gtl_data.zdc(4) <= lhc_data.zdc5g_4;
+    gtl_data.zdc(5) <= lhc_data.zdc5g_5;
+    
 -- HB 2023-02-09: CICADA data mapping
     bjet_l: for i in 0 to NR_BJET_OBJECTS-1 generate
         gtl_data.bjet(i)(BJET_ET_HIGH downto BJET_ET_LOW) <= lhc_data.cicada(i)(10 downto 0); -- bjet Et
@@ -102,7 +110,7 @@ begin
         gtl_data.bjet(i)(BJET_PHI_HIGH downto BJET_PHI_LOW) <= lhc_data.cicada(i)(26 downto 19); -- bjet phi
         gtl_data.bjet(i)(BJET_FLAG_BIT) <= lhc_data.cicada(i)(27); -- bjet flag (currently unused)
     end generate;
-    
+
     gtl_data.cicada_ad(AD_INT_HIGH downto AD_INT_HIGH-3) <= lhc_data.cicada(0)(31 downto 28); -- anomaly detection integer part high
     gtl_data.cicada_ad(AD_INT_LOW+3 downto AD_INT_LOW) <= lhc_data.cicada(1)(31 downto 28); -- anomaly detection integer part low
     gtl_data.cicada_ad(AD_DEC_HIGH downto AD_DEC_HIGH-3) <= lhc_data.cicada(2)(31 downto 28); -- anomaly detection decimal part high