+++ firmware/hdl/payload/gtl_module_tpl.vhd
@@ -2,7 +2,9 @@
 -- Global Trigger Logic module.
 
 -- Version history:
--- HB 2023-06-23: v1.19.0: Implemented logic for CICADA (bjets) on link 11.
+-- HB 2023-09-06: v1.20.0: Implemented logic for CICADA (bjets) on link 11.
+-- HB 2023-09-01: v1.19.1: Added zdc_condition.vhd (was missing) and updated sim and dep file.
+-- HB 2023-08-25: v1.19.0: Used link 71 as ZDC 5G input. ZDC data structure changed (16 bits only on 5G link). Updated top_decl.vhd.
 -- HB 2023-04-15: v1.18.1: Bug fix muon index bits.
 -- HB 2023-03-14: v1.18.0: Implemented "MUS2" (new hadronic shower bit), Anomaly Detection Trigger, ZDC conditions and cuts for muon index bits.
 -- HB 2022-11-29: v1.17.6: Bug fix in algo_pre_scaler_fractional_float.vhd.